Update server.js
In this updated code, the /search-tunes endpoint enables users to search for tunes by querying the title, rhythm, or key fields of the Tune collection. The search query is case-insensitive, thanks to the $regex operator with the $options: 'i' flag.

Additionally, a new function readABCFiles is introduced to read the .abc files from the specified directory and save the tunes' information to the database during the server startup.

Remember to update the abcDirectory variable in the readABCFiles function with the appropriate path to your local directory containing the .abc files.

Please note that this code assumes you have already defined the User and Tune schemas correctly and have the necessary MongoDB setup.